requirements.txt
Flask==3.1.3
Werkzeug==3.1.3
gunicorn==25.1.0
flask-cors==4.0.0

stripe==11.5.0
requests==2.32.5
python-dotenv==1.0.1
faster-whisper==1.2.1
msal==1.31.0
stripe==11.5.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
itsdangerous==2.1.2